requirement.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
cycler==0.10.0
discord.py==1.7.2
idna==3.1
kiwisolver==1.3.1
matplotlib==3.4.2
multidict==5.1.0
numpy==1.20.3
Pillow==8.2.0
ply==3.11
pyasn1==0.4.8
pycryptodomex==3.10.1
pyparsing==2.4.7
pysmi==0.3.4
pysnmp==4.4.12
python-dateutil==2.8.1
python-dotenv==0.17.1
six==1.16.0
typing-extensions==3.10.0.0
yarl==1.6.3
